Only use helper_methods where available
Rails apps generated as Rails API do not include the code for
`helper_method`. This is because helper_method is used in the view
layer, and APIs don't have views.

In discussions about this issue in Rails the opinion was that gems that
use `helper_method` are responsible for fixing the behavior. See:
https://github.com/rails/rails/issues/21067

That makes sense to me.

This fix is a direct lift from

https://github.com/halogenandtoast/monban/pull/47