Require users to opt in to shib stubbing
The current implementation of STUB_INTERNET_ID just checks for the
existence of the ENV variable which can be accidentally set. Change the
`using_stub_internet_id?` to check to see if this kind of stubbing has
been enabled on the system.

The best/simplest solution we found was to put a file in the home
directory (so it is outside the repo). This file can be set on developer
machines, and its absence in deployed environments will prevent the
STUB_INTERNET_ID from being used there.