Update gem, restore test, move to rspec
No functional changes in this commit.

While trying to investigate #16 I tried to run the test. Not tests,
there's only one.

The test had bit-rotted to the point that it was not worth salvaging.
And it was in minitest, which isn't our standard.

Before I fix #16 I'd like to have a test in place, and that required a
working test suite. So I:

- Added RSpec
- Ported the existing test into RSpec
- Configured Rake to run the tests by default
- Added standard Rubocop/Overcommit combo
- Added necessary binstubs
- Added our standard setup/update/test scripts
- Deleted minitest tests