Remove duplicated test
When I first ported this test to RSpec my local copy was way out of
date. The test had already been ported to Rspec in master.

I'm deleting my duplicate test and updating what we use for the
'request_double' in the remaining tests, using a verifiable
double instead of a generic one.