Change behavior of xhr redirect to use root_path
The reason for this change is covered in the addition I made to the
readme.

> In the case of a non-XHR request, shib will redirect the user to their
originally-requested URL. This is fine, as that request was already a
normal HTTP request and should work as expected.

> But in the case of an XHR request we can not redirect the user to the
same URL again. It will not behave the same because the redirect will be
missing the "XMLHttpRequest" header that Rails relies on.

The pre-change code would redirect the user to the modal route, Rails
would see it as an HTML (not JS) request and then render the response as
HTML. Which leads to weirdness.

Sending the user to the Login page & then back to Root seems to be the
path of least surprise here.