Extract redirect_to_shib_login method
Extracting new method `redirect_to_shib_login` method so we have the
option to call this method directly instead of going through the
`shib_umn_auth_required` method to redirect to the shib endpoint when we
do not have a valid shib user.

We created tests for the `redirect_to_shib_login` method by copying them
from `shib_umn_auth_required` method. We left the original tests because
the redirect is still a part of the `shib_umn_auth_required` method.
Removing the duplicated tests is left for a future refactor.