Return emplid and display_name if available
It makes sense to get user information from shibboleth if possible since
it is the source of truth for user data.  This modification will give
access to the display_name and emplid if shibboleth is configured to
give access to that metadata.  See the README file for usage details.

Note that this change required increasing the number of "stub" fields
for testing.  However, the change was backwards compatible with existing
applications that only provide the internet_id stub field.