Allow stubbing in development and test environments
This change removes the dependency on the `~/.umn_shib_auth_enable_stub`
file for enabling internet ID stubbing. Instead, stubbing will
automatically be allowed if the Rails environment is `development` or
`test`. This seems to correspond with the typical use cases for
stubbing; you're either running a server in development mode or you're
running a test suite.

What this doesn't accomodate is folks who use custom environment names
for development. If this causes issues, the list of stubbable
environment could be made configurable.

[Fixes #24]